examples/phone-chatbot/requirements.txt
pipecat-ai[cartesia,deepgram,google,openai,silero]==0.0.67
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
gunicorn==21.2.0
pydantic==2.6.1
typing-extensions==4.9.0
